Scripts/DragonSoul: fixed unary minus applied to unsigned type

+++ src/server/scripts/Cataclysm/Raids/DragonSoul/boss_yorsahj_the_unsleeping.cpp
@@ -987,7 +987,7 @@ public:
                 me->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_DISABLE_MOVE | UNIT_FLAG_NON_ATTACKABLE);
                 me->SetInCombatWithZone();
 
-                if (Unit * target = SelectTarget(SELECT_TARGET_RANDOM, 0, 100.0f, true, -SPELL_FIXATE))
+                if (Unit * target = SelectTarget(SELECT_TARGET_RANDOM, 0, 100.0f, true, -((int32)SPELL_FIXATE)))
                 {
                     me->AddThreat(target, 99999999.0f);
                     me->GetMotionMaster()->MoveChase(target);
